fixed bug with main menu buttons and score not updating

+++ Assets/_Project/Scripts/Score/ScoreHandler.cs
@@ -33,5 +33,5 @@ namespace com.digitalmind.towertest
     }
     
     [Serializable]
-    public class OnScoreChangedEvent : UnityEvent<float>{}
+    public class OnScoreChangedEvent : UnityEvent<int>{}
 }
